requirements.txt
flask==3.1.1
python_dotenv==1.1.1